Add PrintContent Component
