+++ packages/ui/package.json
@@ -4,8 +4,9 @@
   "private": true,
   "exports": {
     "./number-field": "./src/number-field/NumberField.tsx",
+    "./print-content": "./src/print-content/PrintContent.tsx",
     "./button-modal-close": "./src/button-modal-close/ButtonModalClose.tsx",
-    "./pdf-download": "./src/pdf-download/PdfDownload.tsx",
+    "./pdf-result-download": "./src/pdf-result-download/PdfResultDownload.tsx",
     "./modal-building-code-content": "./src/modal-building-code-content/ModalBuildingCodeContent.tsx",
     "./modal-glossary-content": "./src/modal-glossary-content/ModalGlossaryContent.tsx",
     "./walkthrough-card": "./src/walkthrough-card/WalkthroughCard.tsx",
