--- packages/ui/src/pdf-download/PdfDownload.css
@@ -1,218 +1,0 @@
-.ui-pdfDownload-container {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 20px;
-  align-items: center;
-  justify-content: center;
-  background-color: #f9f9f9;
-  border: 1px solid #e0e0e0;
-  border-radius: 8px;
-  max-width: 640px;
-  margin: 0 auto;
-  padding-left: 5px;
-  transition:
-    box-shadow var(--transition-medium),
-    border-color var(--transition-medium);
-}
-
-.ui-pdfDownload-image {
-  border-radius: 8px 0 0 8px;
-  margin: 0 auto;
-  transition: 0.5s ease;
-  backface-visibility: hidden;
-  opacity: 1;
-  height: 100%;
-}
-
-.ui-pdfDownload-imageContainer {
-  position: relative;
-  opacity: 1;
-  transition: 0.5s ease;
-  background-color: #000000;
-  border-radius: 8px;
-  height: 100%;
-}
-
-.ui-pdfDownload-imageMiddle {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #fff;
-  border-radius: 50%;
-  width: 50px;
-  height: 50px;
-  transition: opacity 0.3s ease;
-  opacity: 0;
-}
-
-.ui-pdfDownload-imageContainer:hover .ui-pdfDownload-image {
-  opacity: 0.3;
-}
-
-.ui-pdfDownload-imageContainer:hover .ui-pdfDownload-imageMiddle {
-  opacity: 1;
-}
-
-.ui-pdfDownload-container:hover {
-  cursor: pointer;
-  box-shadow: var(--surface-shadow-medium), var(--surface-shadow-medium);
-  border: var(--walkthrough-card-border-default);
-}
-
-.ui-pdfDownload-title {
-  color: var(--typography-color-primary);
-  font-size: var(--typography-font-size-h5);
-  font-weight: var(--typography-font-weights-bold);
-}
-
-.ui-pdfDownload-button {
-  font-size: 16px;
-  padding: 10px 20px;
-  background-color: #007bff;
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.ui-pdfDownload-button:hover {
-  background-color: #0056b3;
-}
-
-.ui-pdfDownload--DownloadText {
-  padding: 0;
-  gap: 0;
-}
-
-.ui-pdfDownload-aside {
-  display: flex;
-  height: 191px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: var(--layout-margin-large);
-  flex: 1 0 0;
-  width: max-content;
-}
-
-/* Media query for small screens */
-@media (max-width: 980px) {
-  .ui-pdfDownload-container {
-    grid-template-columns: 1fr;
-    gap: 10px;
-    max-width: 400px;
-  }
-
-  .ui-pdfDownload-aside {
-    align-items: center;
-    text-align: center;
-    margin: 0 auto;
-  }
-  .ui-pdfDownload-image {
-    width: 100%;
-    height: 100%;
-  }
-}
-/* Hide the table by default */
-.ui-pdfDownload--printContainer {
-  display: none;
-}
-
-/* Show the table when printing */
-
-@media print {
-  /* Hide everything on the page besides what we need to print */
-  .ui-Header,
-  .web-Walkthrough--StepTracker,
-  .web-WalkthroughFooter,
-  .u-container-walkthrough,
-  .ui-pdfDownload-container,
-  .ui-pdfDownload-button,
-  .ui-pdfDownload-imageContainer,
-  .ui-pdfDownload-aside {
-    display: none;
-  }
-
-  .web-Walkthrough,
-  .web-Walkthrough--Content {
-    display: block !important;
-    min-width: 100%;
-  }
-
-  .web-Walkthrough {
-    height: 100% !important;
-  }
-
-  body {
-    margin: 0;
-    padding: 0;
-    width: 100%;
-    min-width: 100%;
-  }
-
-  .ui-pdfDownload--printContainer {
-    width: 100%;
-    display: block;
-  }
-
-  .ui-pdfDownload--questionTable {
-    display: block;
-    width: 98%;
-    border-collapse: collapse;
-    break-inside: auto;
-    page-break-inside: auto;
-  }
-
-  .ui-pdfDownload--questionTable th,
-  .ui-pdfDownload--questionTable td {
-    border: 1px solid #000;
-    padding: 8px;
-    text-align: left;
-    word-wrap: break-word;
-    font: var(--typography-regular-small-body);
-  }
-
-  .ui-pdfDownload--questionTable th {
-    background-color: #f2f2f2;
-  }
-
-  .ui-pdfDownload--questionTable tr {
-    break-inside: auto;
-    break-after: auto;
-    page-break-inside: auto;
-  }
-
-  .ui-pdfDownload--questionTable td,
-  .ui-pdfDownload--questionTable th {
-    break-inside: avoid;
-  }
-
-  /* Set fixed widths for table columns */
-  .ui-pdfDownload--questionColumn {
-    width: 60%;
-  }
-
-  .ui-pdfDownload--referenceColumn {
-    width: 40%;
-  }
-
-  /* Ensure no overflow and hide any scrollbars */
-  html,
-  body {
-    height: initial !important;
-    overflow: initial !important;
-  }
-
-  .ui-pdfDownload--walkthroughTitle {
-    padding: var(--layout-padding-small);
-    font: var(--typography-bold-h3);
-  }
-
-  .ui-pdfDownload--sectionTitle {
-    font: var(--typography-bold-h5);
-  }
-}
