+++ packages/ui/src/pdf-result-download/PdfResultDownload.css
@@ -1,0 +1,119 @@
+.ui-PdfResultDownload-container {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 20px;
+  align-items: center;
+  justify-content: center;
+  background-color: #f9f9f9;
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+  max-width: 640px;
+  margin: 0 auto;
+  padding-left: 5px;
+  transition:
+    box-shadow var(--transition-medium),
+    border-color var(--transition-medium);
+}
+
+.ui-PdfResultDownload-image {
+  border-radius: 8px 0 0 8px;
+  margin: 0 auto;
+  transition: 0.5s ease;
+  backface-visibility: hidden;
+  opacity: 1;
+  height: 100%;
+}
+
+.ui-PdfResultDownload-imageContainer {
+  position: relative;
+  opacity: 1;
+  transition: 0.5s ease;
+  background-color: #000000;
+  border-radius: 8px;
+  height: 100%;
+}
+
+.ui-PdfResultDownload-imageMiddle {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  color: #fff;
+  border-radius: 50%;
+  width: 50px;
+  height: 50px;
+  transition: opacity 0.3s ease;
+  opacity: 0;
+}
+
+.ui-PdfResultDownload-imageContainer:hover .ui-PdfResultDownload-image {
+  opacity: 0.3;
+}
+
+.ui-PdfResultDownload-imageContainer:hover .ui-PdfResultDownload-imageMiddle {
+  opacity: 1;
+}
+
+.ui-PdfResultDownload-container:hover {
+  cursor: pointer;
+  box-shadow: var(--surface-shadow-medium), var(--surface-shadow-medium);
+  border: var(--walkthrough-card-border-default);
+}
+
+.ui-PdfResultDownload-title {
+  color: var(--typography-color-primary);
+  font-size: var(--typography-font-size-h5);
+  font-weight: var(--typography-font-weights-bold);
+}
+
+.ui-PdfResultDownload-button {
+  font-size: 16px;
+  padding: 10px 20px;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+}
+
+.ui-PdfResultDownload-button:hover {
+  background-color: #0056b3;
+}
+
+.ui-PdfResultDownload--DownloadText {
+  padding: 0;
+  gap: 0;
+}
+
+.ui-PdfResultDownload-aside {
+  display: flex;
+  height: 191px;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: var(--layout-margin-large);
+  flex: 1 0 0;
+  width: max-content;
+}
+
+/* Media query for small screens */
+@media (max-width: 980px) {
+  .ui-PdfResultDownload-container {
+    grid-template-columns: 1fr;
+    gap: 10px;
+    max-width: 400px;
+  }
+
+  .ui-PdfResultDownload-aside {
+    align-items: center;
+    text-align: center;
+    margin: 0 auto;
+  }
+  .ui-PdfResultDownload-image {
+    width: 100%;
+    height: 100%;
+  }
+}
